Mini-Project/Prediction/Mobile_Prediction/requirements.txt
flask
pandas
scikit-learn
joblib
pillow
ultralytics